+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog] and this project adheres to
 
 ## [Unreleased]
 ### Added
+- `PartialOrd` blanket `impl` for `MemOrdered` works for any type that the inner
+  type implements `MemOrd` for.
 - Derive `Clone`, `Copy`, `Debug`, `Default`, and `Hash` for `MemOrdered`
 - `#[inline]` attribute to all `pub` functions
 - Slice conversion methods to `MemOrdered`